+++ src/application/common/events/event.py
@@ -1,0 +1,4 @@
+from typing import Protocol
+
+
+class Event(Protocol): ...
