+++ src/application/common/events/event_bus.py
@@ -1,0 +1,7 @@
+from typing import Protocol
+
+from src.application.common.events.event import Event
+
+
+class EventBus(Protocol):
+    async def publish_event(self, event: Event) -> None: ...
