--- src/application/dinners/events/dummy_handler.py
@@ -1,5 +1,0 @@
-from src.domain.menu.events.menu_created import MenuCreated
-
-
-class DummyHandler:
-    async def handle(self, notification: MenuCreated) -> None: ...
