+++ src/infrastructure/message_broker/message_broker.py
@@ -1,0 +1,1 @@
+class MessageBroker: ...
